chore: remove client code and move server into src
