+++ package.json
@@ -14,7 +14,7 @@
     }
   },
   "bin": {
-    "libp2p-rendezvous-server": "src/server/bin.js"
+    "libp2p-rendezvous-server": "src/bin.js"
   },
   "files": [
     "dist",
