+++ src/bin.js
