+++ src/datastores/interface.ts
