+++ src/errors.js
@@ -1,6 +1,5 @@
 'use strict'
 
 exports.codes = {
-  INVALID_NAMESPACE: 'ERR_INVALID_NAMESPACE',
-  NO_CONNECTED_RENDEZVOUS_SERVERS: 'ERR_NO_CONNECTED_RENDEZVOUS_SERVERS'
+  INVALID_COOKIE: 'ERR_INVALID_COOKIE'
 }
