+++ src/rpc/handlers/unregister.js
