--- src/server/constants.js
@@ -1,13 +1,0 @@
-'use strict'
-
-exports.MAX_NS_LENGTH = 255
-exports.MAX_DISCOVER_LIMIT = 1000
-exports.MAX_PEER_REGISTRATIONS = 1000
-exports.MIN_TTL = 7.2e6
-exports.MAX_TTL = 2.592e+8
-exports.PROTOCOL_MULTICODEC = '/rendezvous/1.0.0'
-exports.GC_BOOT_DELAY = 10e6
-exports.GC_INTERVAL = 7.2e6
-exports.GC_MIN_INTERVAL = 3e6
-exports.GC_MIN_REGISTRATIONS = 1000
-exports.GC_MAX_REGISTRATIONS = 10e6
