--- src/server/errors.js
@@ -1,5 +1,0 @@
-'use strict'
-
-exports.codes = {
-  INVALID_COOKIE: 'ERR_INVALID_COOKIE'
-}
