+++ test/dos-attack-protection.spec.js
@@ -11,10 +11,10 @@ const { toBuffer } = require('it-buffer')
 const multiaddr = require('multiaddr')
 const Libp2p = require('libp2p')
 
-const RendezvousServer = require('../src/server')
+const RendezvousServer = require('../src')
 const {
   PROTOCOL_MULTICODEC
-} = require('../src/server/constants')
+} = require('../src/constants')
 const { Message } = require('../src/proto')
 const MESSAGE_TYPE = Message.MessageType
 const RESPONSE_STATUS = Message.ResponseStatus
